rest-api/requirements.txt
fastapi
uvicorn
mysql-connector-python
confluent_kafka
msgpack